fixed a typo in the OSAtomic_mac.cpp

+++ clang/test/Analysis/OSAtomic_mac.cpp
@@ -33,7 +33,7 @@ void gh197211_test() {
 
 bool OSAtomicCompareAndSwapMismatchedPointee(int __oldValue, int __newValue,
                                              volatile long *_theValue);
-int gh197211_long_flag;
+long gh197211_long_flag;
 void gh197211_test2() {
   if (OSAtomicCompareAndSwapMismatchedPointee(0, 0, &gh197211_long_flag))
     ;
